[Backend] Refactored route to get force's patents
